refactor(GCS+gRPC): subdirectory for async implementation
