+++ google/cloud/storage/google_cloud_cpp_storage_grpc.bzl
@@ -20,9 +20,9 @@ google_cloud_cpp_storage_grpc_hdrs = [
     "async_client.h",
     "async_object_responses.h",
     "grpc_plugin.h",
-    "internal/async_accumulate_read_object.h",
-    "internal/async_connection.h",
-    "internal/async_connection_impl.h",
+    "internal/async/accumulate_read_object.h",
+    "internal/async/connection.h",
+    "internal/async/connection_impl.h",
     "internal/grpc/bucket_access_control_parser.h",
     "internal/grpc/bucket_metadata_parser.h",
     "internal/grpc/bucket_name.h",
@@ -58,8 +58,8 @@ google_cloud_cpp_storage_grpc_hdrs = [
 google_cloud_cpp_storage_grpc_srcs = [
     "async_client.cc",
     "grpc_plugin.cc",
-    "internal/async_accumulate_read_object.cc",
-    "internal/async_connection_impl.cc",
+    "internal/async/accumulate_read_object.cc",
+    "internal/async/connection_impl.cc",
     "internal/grpc/bucket_access_control_parser.cc",
     "internal/grpc/bucket_metadata_parser.cc",
     "internal/grpc/bucket_name.cc",
