+++ google/cloud/storage/google_cloud_cpp_storage_grpc.cmake
@@ -36,11 +36,11 @@ else ()
         async_object_responses.h
         grpc_plugin.cc
         grpc_plugin.h
-        internal/async_accumulate_read_object.cc
-        internal/async_accumulate_read_object.h
-        internal/async_connection.h
-        internal/async_connection_impl.cc
-        internal/async_connection_impl.h
+        internal/async/accumulate_read_object.cc
+        internal/async/accumulate_read_object.h
+        internal/async/connection.h
+        internal/async/connection_impl.cc
+        internal/async/connection_impl.h
         internal/grpc/bucket_access_control_parser.cc
         internal/grpc/bucket_access_control_parser.h
         internal/grpc/bucket_metadata_parser.cc
@@ -186,8 +186,8 @@ if (BUILD_TESTING AND GOOGLE_CLOUD_CPP_STORAGE_ENABLE_GRPC)
     set(storage_client_grpc_unit_tests
         # cmake-format: sort
         grpc_plugin_test.cc
-        internal/async_accumulate_read_object_test.cc
-        internal/async_connection_impl_test.cc
+        internal/async/accumulate_read_object_test.cc
+        internal/async/connection_impl_test.cc
         internal/grpc/bucket_access_control_parser_test.cc
         internal/grpc/bucket_metadata_parser_test.cc
         internal/grpc/bucket_name_test.cc
