+++ google/cloud/storage/internal/async/connection.h
